api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Modular RAG API with JWT authentication"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core API
    "fastapi>=0.110.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.0",
    # Databases
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "dnspython>=2.6.1",
    # AI / RAG
    "pinecone>=5.0.0",
    "numpy>=1.25.0",
    "google-genai>=1.49.0",
    # File processing
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "beautifulsoup4>=4.14.2",
    "markdown>=3.10",
    "reportlab>=4.0.0",
    "bs4>=0.0.2",
    "sarvamai>=0.1.22",
    "fastembed>=0.2.0",
    "flashrank>=0.2.10",
]
